Rename class DifferentiableFunctionStack -> DifferentiableFunctionSet.
  - make old name deprecated.
