+++ src/CMakeLists.txt
@@ -16,7 +16,7 @@
 SET(LIBRARY_NAME ${PROJECT_NAME})
 SET (${LIBRARY_NAME}_SOURCES
   differentiable-function.cc
-  differentiable-function-stack.cc
+  differentiable-function-set.cc
   generic-transformation.cc
   relative-com.cc
   com-between-feet.cc
